Create useHomeData hook
